+++ js/lib/validate.ts
@@ -1,18 +1,43 @@
 import toml from "@iarna/toml";
-import Ajv, { ValidateFunction } from "ajv";
+import Ajv, { AsyncSchema, AsyncValidateFunction } from "ajv";
 import betterAjvErrors from "better-ajv-errors";
 import fs from "fs/promises";
 import _ from "lodash";
 import path from "path";
 
-// TODO: can we validate Tracing tests?
-// how to add a format to ajv
+import type { Question } from "./questions/mod";
+import { questionValidators } from "./questions/validate";
 
 export class Validator {
-  constructor(readonly schema: any, readonly validator: ValidateFunction) {}
+  constructor(readonly schema: any, readonly validator: AsyncValidateFunction<void>) {}
 
   static async load(distDir: string): Promise<Validator> {
     let ajv = new Ajv();
+
+    ajv.addKeyword({
+      keyword: "questionType",
+      async: true,
+      async validate(questionType: Question["type"], question: Question) {
+        if (process.env.QUIZ_LIGHTWEIGHT_VALIDATE !== undefined) {
+          return true;
+        }
+
+        let validator = questionValidators[questionType];
+        if (validator) {
+          let message = await validator(question.prompt, question.answer);
+          if (message) {
+            throw new Ajv.ValidationError([
+              {
+                keyword: "questionType",
+                message,
+              },
+            ]);
+          }
+        }
+        return true;
+      },
+    });
+
     ajv.addFormat("markdown", (_data: string) => {
       // TODO: how should we validate Markdown? Print it to the console for visual inspection?
       return true;
@@ -20,22 +45,30 @@ export class Validator {
 
     let schemaPath = path.join(distDir, "Quiz.schema.json");
     let schemaRaw = await fs.readFile(schemaPath, "utf-8");
-    let schema = JSON.parse(schemaRaw);
-    let validator = ajv.compile(schema);
+    let schema: AsyncSchema = JSON.parse(schemaRaw);
+    let validator = ajv.compile<void>(schema);
 
     return new Validator(schema, validator);
   }
 
-  validate(input: string): string | undefined {
+  async validate(input: string, quizPath: string): Promise<string | undefined> {
+    let quiz = toml.parse(input);
     try {
-      let quiz = toml.parse(input);
-      if (!this.validator(quiz)) {
-        return betterAjvErrors(this.schema, quiz, this.validator.errors!, {
-          indent: 2,
-        });
+      await this.validator(quiz);
+    } catch (err: any) {
+      if (err instanceof Ajv.ValidationError) {
+        let formattedErrors = betterAjvErrors(
+          this.schema,
+          quiz,
+          err.errors.filter(err => !["const", "anyOf"].includes(err.keyword!)) as any,
+          {
+            indent: 2,
+          }
+        );
+        return `Invalid quiz: ${quizPath}\n${formattedErrors}`;
+      } else {
+        return err.toString();
       }
-    } catch (e: any) {
-      return e.toString();
     }
   }
 }